Create BMS C# Code
Building management system is a system that control electronic and mechanical devices in the building through group of controllers and sensors that is spread in the building.
The system provides many features such as the automated services which enable users to use it more easily, more over it created more convenient environment for users. It also provides the ability of monitoring and controlling on these devices, and the ability to know the amount of power that has been used which will be done by central computers. Those central computers are controlled by administrators who can control and manage floors of building through multi-interfaces that display every part in the building. These multi-interfaces provide many controlling features which will enable administrators to control the hole building.  
The idea of BMS was created to cover the lack of smart buildings due to technology improvement. These buildings will provide many services needed nowadays, and they will save use of power which will help in protecting environment. We are expecting the system to provide comfort as well as saving time and effort.